Re-organize example scripts.
