--- Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/ISystem.cs
@@ -1,4 +1,0 @@
-﻿public interface ISystem
-{
-    void Perform();
-}
