--- Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/NativeSystem.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-public class NativeSystem : ISystem
-{
-    public void Perform()
-    {
-        Debug.Log("The native system is performing its functionality.");
-    }
-}
