--- Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/NativeSystemGlobalContainerWrapper.cs
@@ -1,20 +1,0 @@
-﻿using Chopsticks.Dependencies.Containers;
-using Chopsticks.Dependencies.Services;
-using UnityEngine;
-
-public class NativeSystemGlobalContainerWrapper : MonoBehaviour
-{
-    private DependencyRegistration _registration;
-
-
-    public void OnEnable()
-    {
-        MonoContainerService.GlobalContainer.Register(
-            new NativeSystem(), out _registration);
-    }
-
-    public void OnDisable()
-    {
-        MonoContainerService.GlobalContainer.Deregister(_registration);
-    }
-}
