+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e24537813dce3d45aa21bc3fa793a5d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
