+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/Counting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17f3accb01421274db43def5fe782345
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
