+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/Counting/UnalterableCountingSystemSuperclass.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 66dd395a4833c34488585853fddebca5
