+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/ISystem.cs
@@ -1,0 +1,8 @@
+﻿
+namespace Chopsticks.Samples.ExampleDependencies.System
+{
+    public interface ISystem
+    {
+        void Perform();
+    }
+}
