+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/Native.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc2b853d56f5ef546aa1a6571dbd2595
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
