+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/Native/NativeSystem.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace Chopsticks.Samples.ExampleDependencies.System.Native
+{
+    public class NativeSystem : ISystem
+    {
+        public void Perform()
+        {
+            Debug.Log("The native system is performing its functionality.");
+        }
+    }
+}
