+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/Native/NativeSystemGlobalContainerWrapper.cs.meta
