+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependencies/System/Output.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74b87127adf153746908989538970d6e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
