+++ Unity/com.chopsticks.dependencies/Assets/Samples/Scripts/ExampleDependents/InputHandler.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe694176cfa713d4f833ed57339c605c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
